refactor: relocate more integration specific utility methods
