+++ src/integrations/Iterable/utils.js
@@ -1,4 +1,4 @@
-import { getDataFromSource } from '../../utils/utils';
+import { getDataFromSource } from '../../utils/commonUtils';
 
 const ITEMS_MAPPING = [
   { src: 'product_id', dest: 'id' },
