+++ src/integrations/Sentry/browser.js
@@ -6,8 +6,7 @@
 import get from 'get-value';
 import logger from '../../utils/logUtil';
 import { SentryScriptLoader, sentryInit } from './utils';
-import { removeUndefinedAndNullValues, getDefinedTraits } from '../../utils/commonUtils';
-import { isObject } from '../../utils/utils';
+import { removeUndefinedAndNullValues, getDefinedTraits, isObject } from '../../utils/commonUtils';
 import { NAME } from './constants';
 
 class Sentry {
